currently on a much more positive trajectory than last.
I have stopped with all the silly design fuss, i fixed the navigation system so that it doesn't do weird unexpected things.

And i have now made a good platform for adding the core functionality of the app, that being the ability to edit the json in a safe manner, as well as being able to preview the links to filter them out.

so i think in a short manner it will be possible to get the core functionality working, and from there adding more polish and such.

yeah, it surely can't be a long time until i get to move onto the other stuff, that being scraping the actual articles, which might require its own one of these...