added a function that create's the cache folder, as well as tidying up in the python code
I added the function to create the cache folder, because it doesn't normally hold any data, apart from when it's being run (the script linkScraperGui), so therefore there aren't any files in it.

But git only tracks file names, and not folders, so when i git pushed or something cache wouldn't be stored, but by having it be made by the script upon startup that  problem dissapears, and is no longer a problem.

And then I just removed imports which weren't being used, in an attempt to speed up loadtiems